+++ src/soc/mediatek/common/mt6363_sdmadc.c
@@ -61,11 +61,11 @@ static const struct auxadc_chan_spec mt6363_sdmadc_chan_specs[] = {
 _Static_assert(ARRAY_SIZE(mt6363_sdmadc_chan_specs) == AUXADC_CHAN_MAX,
 	       "Wrong array size for mt6363_sdmadc_chan_specs");
 
-int mt6363_sdmadc_read(enum auxadc_channel channel, int *val,
+int mt6363_sdmadc_read(enum auxadc_channel channel, u32 *val,
 		       enum sdmadc_pures pures, enum auxadc_val_type type)
 {
 	u8 wdata;
-	u32 regval;
+	u16 regval;
 	u32 elapsed = 0;
 	const struct auxadc_chan_spec *chan;
